Issue #224: make fails on Suse 9.1
Patch contributed by Wolfgang Fritz.

* src/rapidsvn_frame.cpp
  (RapidSvnFrame::RapidSvnFrame): explicit type cast
    to "long int" for pConfig->Read

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7101 65390229-12b7-0310-b90b-f21a5aa7ec8e

+++ src/rapidsvn_frame.cpp
@@ -624,7 +624,7 @@ RapidSvnFrame::RapidSvnFrame (const wxString & title)
   m_info_panel->SetSizer (sizer);
 
   // Read frame position
-  if (pConfig->Read (ConfigMaximized, (int)0) == 1)
+  if (pConfig->Read (ConfigMaximized, (long int)0) == 1)
     Maximize (true);
   else
   {
